backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
sqlalchemy==2.0.28
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.27.0
